Construct build rules
I am starting with a collection of build rules that includes some
hooks for a larger "recursive build" facility, as I suspect that
this simulation framework might some day grow to need it,
and wiring in some things now will help reduce the impact.

I like to organize my files to make it easy to look at all the
sources, or all the headers, in a project (or sub-project). Thus
the hoop jumping in the GNUmakefile.rules with settting up
and using directory name variables.