Achieve execution of NOP instructions
Aim at a class method in Cpu8080impl for the logic executed
in each state (see MCS diagram at end of section 2).

Each function checks where in the state we are, does the
logic, and assures that data are set up to get us into the
next state properly.

For now, the T3 state latches the data into IR but always
dispatches to the T4 state for the NOP instruction.

Also:
- turn on -O3 optimizer
- format the source with clang-format
- stub.{h,hh} provides my usual STUB messages
- break tau.h out of tau.hh