Edges have Names, and use the [1] Trick
Being able to print what is going on with edge names and
service function name and parameter names is useful.

Do it now when the change impacts very little code.

Also, I can eliminate a lot of "*" and "&" syntactic noise
by making use of a trick, defining a type to be an array
of one item, with a lesser-used type that is explicitly
a pointer to the item.