Set up for external automatic build
Set up a "make loop" target for a build I can run in a loop
that rebuilds what needs rebuilding and runs tests, which
can be repeatedly run in a window separate from the editor.

Set up a "make await" target that pauses until a file used in
the build is modified (or the target is deleted).