Update README
- Document commitment to PROJECT DOWNCODE
- document the various code modules present

start to draw a distinction between core modules that
are expected to carry over as we simulate different things,
from modules that are specific to a simulation.

For example, we now have "Gen8224" which simulates
a chip very specific to the Intel 8080. It is unlikely this
module will be needed if, in the future, we simulate a
wildly different system.