Refactor Cpu8080_state
Change the type from "function" to "pointer to function"
to get rid of a lot of syntactic noise.

Also, provide the M1T1 state in Cpu8080 so the implementations
of the state functions that are not in cpu8080.c can properly
return control to M1T1 when their instructions are complete.