Start work on interrupts
Anticipating asynchronous arrival of INT which needs
to change the next M1T1 state; modify the existing
    copy cpu->M1T1 into cpu->state_next
because the interrupt could arrive after that is done.
Instead, introduce the "RETURN M1" signal that only
appears in one 8080 timing diagram -- active during
the last T1 of an instruction -- and use that signal
during PHI1 rising to decide to switch to M1T1 rather
than state_next.

Update DI and EI to manipulate CPU->INTE signal

Add INT_RST representing the synchronized reset; this
is not latched by phi1, but instead is directly manipulated
by the state functions related to reset.

Also:
- add "disassemble" method to improve log output
- Add executions of EI and DI to Cpu8080_bist
- add new signals to the Cpu8080_bist trace outputs
- WAIT was not being asserted during M1 TW cycles.