A bit of refactoring
Warnings about switches over enums not covering all
enumeration values are usually good, because they remind
to consider what to do in each place when you add an
enum value. In the case of clock phase: if a phase is
added, we want existing switch code to do nothing.

So drop to using an int for phase, and defined values;
and sweep away all the empty switch cases.

Other big change is to put the static functions into a
more readable order. The only cost is (GASP!) having
to have forward declarations for them.

- Clear the RETM1 signal when it is consumed.
- YAGNI the INT_RST signal.
- provide BUS_FLOAT to use when releasing a bus
- accept modified test results with updated timing