Demonstrate the INT cycle with EI and NOP
RST is not yet implemented, but I can trigger interrupts
and feed in EI and NOP instructions from the device.

Verify that one more instruction after EI is executed
before the interrupt is taken, as part of the test that
it fetches properly.

As part of the comparison, we also can see that PC
is not incrementing when we fetch the opcode from
the interrupting device.