fix includes in "make format"
C files called foo.c:
  #include "foo.h"
  <blank line>
  #include <...> // all system includes, sorted
  <blank line>
  #include "..." // all local includes, sorted
  <blank line>
  <the rest of the file>

H files:
  #include <...> // all system includes, sorted
  <blank line>
  #include "..." // all local includes, sorted
  <blank line>
  <the rest of the file>

Some sections may be blank. Multiple blank lines
are reduced to one blank line. If the C file does not
have its own H file, then also drop the blank line at
the top of the file.

This also means that multiple blank lines in a row
in the source file collapse to single blank lines; this
is a feature, not a bug.